use VarInt as packet id

+++ src/main/java/com/github/mkorman9/game/dto/packet/Response.java
@@ -1,5 +1,5 @@
 package com.github.mkorman9.game.dto.packet;
 
 public interface Response {
-    short packetId();
+    int packetId();
 }
